backend/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.42.0
anyio==4.8.0
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
defusedxml==0.7.1
distro==1.9.0
ecdsa==0.19.0
fastapi==0.115.6
fonttools==4.55.3
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.8
langchain==0.3.14
langchain-anthropic==0.3.1
langchain-core==0.3.29
langchain-text-splitters==0.3.5
langsmith==0.2.10
matplotlib==3.10.0
multidict==6.1.0
numpy==2.2.1
openai>=1.0.0,<2.0.0
orjson==3.10.13
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pillow==11.1.0
pluggy==1.5.0
propcache==0.2.1
pyasn1==0.6.1
pydantic==2.10.4
pydantic_core==2.27.2
pyparsing==3.2.1
pytest==8.3.4
pytest-asyncio==0.25.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
seaborn==0.13.2
sentry-sdk==2.19.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
tenacity==9.0.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.34.0
yarl==1.18.3
streamlit>=1.24.0
plotly>=5.13.1
